+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- `Innmind\Validation\Failure::of()` is now longer flagged as internal
+
 ## 2.0.0 - 2025-04-12
 
 ### Added